[TASK] Remove mod.web_layout.tt_content.fieldOrder
fieldOrder was used to re-arrange specific fields in quick edit mode of the page module.
This however is mostly broken since 4.5 with all the --div-- and --palette-- containers and
only a couple of elements can be moved around successfully. This option is now removed
from documentation in 6.2 and master and will be dropped code wise in master completely.