Drop linkTitleToSelf
TCEFORM."table"."field".linkTitleToSelf can be used in FormEngine to link the
label of a single rendered field to a "sub form" of the displayed row where only
this single field is displayed. This has zero advantage usability wise.
Documentation of this feature is now removed in 6.2 and master, and it will
be dropped code wise in master completely.