Fix formatting which resulted in blockquotes
- fixed indenting
- removed extra newline after :aspect:`Example`

The indenting on the affected pages was also changed to use only
3 spaces indenting (as is defined in .editorconfig). There was
a mixture of 3 and 4 spaces indenting which makes it very difficult
to maintain the pages properly.

Since a lot of text was changed, it will probably not be possible
to backport, suggest to create a separate page for the other branches
or leave it be for now.

Related: TYPO3-Documentation/T3DocTeam#150
Related: TYPO3-Documentation/T3DocTeam#191