Add compatibility section for using page.tsconfig to v11 as well
Since TYPO3 v12, the file Configuation/page.tsconfig ist automatically
loaded. In order to achieve this in a compatible way which works
with v11 and v12 and requires not further changes when updating to
v12, a section was added to main / v12 branch.

However, this is missing in v11 version (where it is even more useful)
and this results in a broken link in TYPO3 Explained (see
TYPO3-Documentation/TYPO3CMS-Reference-CoreApi#2784).

Resolves: TYPO3-Documentation/TYPO3CMS-Reference-CoreApi#2784